Support window groups (henceforth known as gangs) and modal dailogs.
Window gangs are defined by clients using WM_CLIENT_LEADER and
WM_HINS.group_leader properties. A client will often have a main window and its
subwindows belonging to the same gang. The (window-gang) function may be used
to return a list of other members of a given window's gang.

Modal windows and modal dialogs have the property of blocking input to the
application for the life of the window. Without WM support, this can lead to
the confusing situation wherein a program appears to be frozen because its
modal dialog has become obscured. Therefore, modal dialogs will always be kept
above the windows in their 'shadow' (usually includes other dialogs of the
application.) Additionally, moving a modal window to another group will cause
the windows in its shadow to be moved. Likewise, moving a shadowed window will
cause the modal shadowing it and all the other windows in the modal's shadow to
be moved. This is to ensure that a modal window and frozen, shadowed windows
will never belong to different groups, which could be very confusing for the
user.