Fix :sloppy mouse focus for group switching and dialogs.
This and 41b5a67e42c2a42b8e1560c65fedbde3e6235c25 cause the :enter-notify
handler to ignore events for windows not (frame-window) of some frame in the
current groups of all screens. That is, windows not on top. This fixes the bug
where switching groups with :sloppy enabled caused an infinite loop of
refocusing, as well as the bug where it was impossible to focus a fixed-size
window (the window under it being focused when the mouse passed over).