change the way input is handled
No longer call grab-keyboard when stumpwm wants to read input. When a
top level key is pressed, the keyboard is frozen. When a key is to be
read the keyboard is unfrozen for 1 keystroke and then refrozen. When
the key press event has been handled, stumpwm unfreezes the keyboard.

I still don't understand why C-t ; reload lets you type in the focused
window. The keyboard should be frozen at this point.