Documentation improvements
- Created HACKING with some information about StumpWM-specific code
such as defcommand, as well as some general pointers about working on
StumpWM that may seem like common sense and a section on effectively
using git.

- Edited README a little bit, added info. on using asdf-install to
install cl-ppcre on clisp.

- Um, added self to AUTHORS (I think that's the protocol...feel free
to undo if not; did this less out of desire for credit than so
questions RE: contributions could be directed more easily) and lined
up the columns a bit.