Remove :maxsize from window types.
If a window qualified as :maxsize, this fact would cause other, more
important types (such as :dock) to be masked. Ideally we should
support lists of types, as X does. For now, there's
(xwin-maxsize-p). This change allows StumpWM to detect more panels
such as fbpanel and xfce4-panel in addition to gnome-panel and kicker.