workaround the clx pixmap/window error
In both the clisp and sbcl clx libraries, sometimes in te event slots,
what should be a window will be a pixmap instead. In this case, we
need to manually translate it to a window to avoid breakage in
stumpwm. So far the only slot that seems to be affected is the :window
slot for configure-request and reparent-notify events. It appears as
though the hash table of XIDs and clx structures gets out of sync with
X or perhaps X assigns a duplicate ID for a pixmap and a window.

Since I haven't found a way to reliably reproduce this bug, this code
is untested.