Tweak neighbour to work more intuitively.
Simple demonstration of the effect this has:
1. hsplit your screen so you have 2 columns.
2. vsplit one of the columns. You'll have something like this:
|---|---|
|   |   |
|---|   |
|   |   |
|---|---|
3. From the frame on the right, run (move-focus :left)

Before, you end up on the bottom left frame. With this, you end up at
the top left. Similarly, when moving :down or :up to frames that have
equal width, you'll now select the left-most instead of the
right-most. This makes more sense to me.

 frame.lisp |    2 +-
 1 files changed, 1 insertions(+), 1 deletions(-)