change kmaps from a hashtable to a struct containing a list
This change was needed because there are now multiple top level maps
that all hang bindings off the escape key. set-prefix-key cannot know
what top level maps it needs to search to update the escape key, so
*escape-key* was created which is a key struct that is used in any map
that wants to use the escape key. Now that key's slots just need to be
updated. But a hashtable's keys cannot change or the lookup fails, so
a list is now used.