Make gnome-panel recognize closed windows
My version of gnome-panel ignores _NET_CLIENT_LIST updates when they
are not consistent with the _NET_CLIENT_LIST_STACKING state.  However,
when a window was unmapped, stumpwm was previously doing the following:

1. update _NET_CLIENT_LIST property based on `screen-mapped-windows', which
   does not include the newly-unmapped window.
2. update _NET_CLIENT_LIST_STACKING property based on `all-windows', which is
   created by concatenating all the windows in the group.

Problem: `withdraw-window' was calling `remove-mapped-window' (which sends the
updates) BEFORE it removed the unmapped window from its tile-group.  So the
unmapped window was removed from _NET_CLIENT_LIST, but not from
_NET_CLIENT_LIST_STACKING.  But gnome-panel ignores client list updates
whenever _NET_CLIENT_LIST and_NET_CLIENT_LIST_STACKING are inconsistent.

So the fix, which is much shorter than this comment, changes `withdraw-window'
to remove the withdrawn window from its tile-group and then call
`remove-mapped-window'.