Fix scary exchange-direction error messages.
Trying to exchange-direction (1) when no window is in the focused
frame, (2) in a direction with no windows, or (3) off the edge of the
screen causes an ugly error message.

So, with this frame setup:
┌────────┬─────────┐
│(empty) │ (window)│
│        │         │
│        │ :)      │
│        │     =D  │
│        │         │
│        │ bobf<3  │
│        │         │
└────────┴─────────┘

Trying to exchange-direction in any direction with either frame
selected causes an error message (3 different ones depending on the
action, too!), for a total of 8 ways to get yelled at :).

PROOF OF FEROCIOUS WILL