Don't dereference symlinks in program completion list.
The directory function dereferences symlinks in its results, so if you
have /usr/bin/firefox -> ../lib/iceweasel/iceweasel (as debian does),
you won't be able to complete on firefox.  This patch fixes that.
I've included support for as many lisp implementations as I could, but
I've only tested this end-to-end on sbcl, clisp and ccl.