Tidied up run-prog and run-prog-collect-output.
These functions were getting into a bit of a state, so I've had a tidy
up.  The result is by no means perfect, but I think it's an
improvement.  I've removed support for a couple of ancient versions
(CLISP before version 2.26 and Lucid before it became Liquid), and
added support for setting DISPLAY to the current screen for more
lisps.  As an added bonus this patch fixes the build on ECL (at least
from a REPL).