Fixing make-image.lisp for CCL
CCL:SAVE-APPLICATION by default just dumps a heap image, which is
used with the CCL kernel (this kernel/image division is found in
other lisps too, such as SBCL). This image can't be run without
the separate kernel file.

To get CCL to dump an executable image, you can request that the
kernel be prepended to the image.