OpenBSD support for contrib/battery-portable.lisp
I wrote some quick openbsd support for the battery-portable
contrib. Tested on a single-battery laptop and an (obviously)
no-battery desktop.

Note that the change to the fmt-time function is necessary because
openbsd doesn't report a battery life estimate when the battery is
charging, the time will be nil in that case.