Right-justification of mode line elements
Hello list,

attached is a patch which allows for text elements of the mode line
(and those of the message window, if that is required) to be
right-justified.

For this to work you have to put a '^>' into the respective format
command, for instance:

(setf *screen-mode-line-format*
      (list "[%n:%h] %v^>"
            '(:eval (run-shell-command "echo -n `date +'%a %e-%b %R'`"
	          t))))

It also works in case of multiple lines of text. Everything after '^>'
up to the end of the respective line is right-justified. The next line
then starts left-justified again.

Nils