Add :_NET_WM_DESKTOP to +netwm-supported+ to make OpenOffice.org happy.
That seems to make OpenOffice.org behave correctly, when it's missing
it keeps sending client-messages requesting group switches when you
move it in another group. It may improve the stumpwm experience with
other java software as well.

I've also added a call to force-output in dformat, I think it's better
not to buffer writes to *debug-stream* in case of a crash, it also
make it easier to find out what's happening in "real time" when you
tail -f the log file.