Fixed 'STUMPWM' not foun when trying to build
Added STUMPWM_ASDF_DIR in configure.ac to store the current directory
where stumpwm is being configured. This variable is used on
`load-stumpwm.lisp.in`, a new file created with the sole purpose of
bootstrap the loading of `stumpwm.asd`, using ASDF3 config API.

Both `Makefile.in` and `make-image.lisp.in` were modified to
accommodate this change.