Adding Window Selection Expressions
Window Selection Expressions (WSE) are calls to the function
act-on-matching-windows as defined in wse.lisp. The idea is declarative:
you specify how to distinguish the windows you need and your code will
get called on all the matching windows in th specified scope.

In addition to the main function, short helpers for typical predicates
and actions (like pulling window ti the current frame or checking
whether window title matches a regular expression are provided.