Use exclusively ASDF to load StumpWM modules
A StumpWM module is a loadable ASDF system.  Typically, it consists of
an `.asd' file (which defines an ASDF system) and one or more Lisp
source files.

By using ASDF, StumpWM doesn't need to worry how to look for files or
resolve dependencies.  This is all done by ASDF.

This patch aims at ASDF 3.0, which was released on January 31st 2013.

For backward compatibility, the function `load-module' was kept, so
code that uses it continues to work, but when used as a command, it
does not complete on module names.

By getting rid of `*contrib-dir*' and `*load-path*', the user has to
instruct ASDF where it should look for `.asd' files, as they would for
any other Common Lisp system.