Rework color handling and drawing
This replaces the two existing color parsers with a single, different
one. This parser has the same functionality as the parser previously
used by PARSE-COLOR-STRING, in that it turns a string with color codes
into a list of strings and lists of "modifiers", such as (:fg 1) to set
the foreground color to color 1. This parser, however, also supports
including those lists directly in the string. For example:

"Hello ^(:fg 1)World"

would result in a list like '("Hello " (:fg 1) "World").

The rendering code now also uses the output of this new parser, instead
of having its own. The modifiers are passed through the generic function
APPLY-COLOR, which has methods for all of the existing modifiers.
The current rendering state, such as the colors, whether brightness is
enabled, etc, is now stored in the color context object. The different
methods on APPLY-COLOR apply their changes to the instance of CCONTEXT.