Add multiple font support
This adds a ^f<n> modifier to switch to the font at N in the screen's
list of fonts. SET-FONT is changed to support specifying a list of fonts
instead of just one.

As an example, it's now possible to place this in your .stumpwmrc:

(set-font '("9x15bold"
            "-*-comic sans ms-medium-*-*-*-*-*-*-*-*-*-*-*"))

Then you could set *STARTUP-MESSAGE* to "Hello ^f1Comic Sans", which
would result in "Hello " being rendered in the default font, and "Comic
Sans" being rendered in Comic Sans MS. The height of each line of output
is determined by the tallest font used on it, and the rendering code
will try to center each piece of text on a line vertically.