Stop generating stumpwm.texi using lisp
Instead of generating the source of the info manual from a mixture of
stumpwm.texi.in and documentation scrapped from lisp source files, use
a single manual source, stumpwm.texi.

stumpwm.texi was first generated using the function `generate-manual'
found in `manual.lisp', which has since been deleted.

The expressions in `Makefile.in' used by each lisp implementation to
generate the .texi file have been deleted and a new make rule, `info',
has been added.

The removal of the generation step reduces the complexity of making
the manual because the manual source file being edited is the only
source for the Info manual.  It is not necessary anymore to maintain
specific expressions for each lisp implementation supported to
generate the manual source file.

Docstrings and manual pages will start to drift apart because they are
not automatically in sync anymore, but this is a good thing as they
have different purposes.  This unfortunately imposes a burden to the
project because now there are two places to describe sometimes similar
things.