more tolerant window-head (float-window)
We return the head that contains one of the top-left, top-right,
bottom-left, bottom-right window corners.  If no such head is found, we
just return the first head in the screen's heads list - it's better than
returning NIL.

The old way couldn't deal with windows that are partly out of the frame.