Sane behavior of the sloppy mouse focus policy in floating groups.
The behavior of *mouse-focus-policy* = :sloppy in floating groups is:
any window that gets the focus (e.g. by hovering over it with the
mouse) automatically raises. This is annoying and makes using floating
groups very difficult. This patch makes floating group windows grab
the focus but stay in the background.  Windows raise only when
clicked.