Implement StumpUI for internal windows
It defines some protocols that abstract window creation and rendering
away, allowing the places that currently do it themselves to be cleaned
up.

This is mostly based on how the message window is currently handled,
with some added flexibility.

Before everything can be ported over, there will also need to be a
protocol for input handling.