Rename &float-focus-next to %float-focus-next
The & sigil is used to refer to lambda list keywords. It confuses Emacs's
syntax highlighting, which highlights calls to the function as lambda list
keywords even outside. Replace it with the more common idiom of using the % to
signal that it is an internal function that shouldn't be called directly.