Handle user-provided declare forms correctly in event-handlers
The expansion of DEFINE-STUMP-EVENT-HANDLER would place any extra declarations
provided by the user of the macro below the always present (declare (ignore
,event-slots)) forsm. So any extra declare forms provided by the user of the
macro would be discarded as a there can only be one declare expression at the
start of the function body. Merge the user-provided declarations with the first
declare form.