Use UIOP for pathname handling instead of cl-fad extract
UIOP is available by default in SBCL due to SBCL bundling asdf as a
dependency. As UIOP is meant to be a wrapper layer on top of different
lisp distributions and is bundled in the most well used one it could
be considered a almost standard pathname api for modern systems. Due
to this using UIOP for pathnames instead of cl-fad allows us to reduce
our maintenance burden through using a proper library.