Store frames as binary trees and make BALANCE-FRAMES work as in Emacs
This patch does two related things:

1. Changes frame trees to be stored as binary trees instead of plain old
   vanilla n-ary trees.
2. Changes balance-frames to match Emacs' balance-frames and fully
   balance all frames.

Part 1 is what was done in https://github.com/stumpwm/stumpwm/pull/447
which was reverted later because it broke frame balancing.  The upside
to storing frames in a binary tree is described in that PR, but in
short: it means that if you've got your frames in some layout, then
split a frame, and then immediately delete it, you'll end up back in
your original layout.

Part 2 fixes the frame balancing to work with the new binary tree
scheme.  It makes the `balance-frames` command match the one in Emacs,
which balances *all* frames, not just one particular level.  If this
isn't desireable and the old level of only balancing some frames is
wanted, it should be possible to modify this patch to account for that.
This is mostly just a starting point.

The actual balancing algorithm is a bit wordy because it has to handle
the two split cases separately.  Counting the number of side-by-side
splits of a given type is also slightly trickier because it's no longer
a single list within the tree, but it's not too bad to do it
recursively.

I'm not 100% sure this works properly with multiple montior setups.
I'll have to test it on my other setup (which has two separate monitors)
when I get a chance.