Added support for a new completion style.
The new style, instead of cycling through possible completions,
only completes when there is exactly one match, otherwise it displays
a message window with the possible completions.

Added special variable *input-cycle-completions* that enables
the new completion style when its value is nil.

Added special variable *message-window-input-gravity*
that is to be used as a message window's gravity
when the input window is being displayed so that they do not overlap.

The function message-window-real-gravity should be used to compute
the proper gravity to use when displaying a message window.