New message window gravity setting
The same gravity setting was being used for both the message window
and the input window, making displaying both at once problematic.

Added special variable *message-window-input-gravity*
that is to be used as a message window's gravity
when the input window is being displayed so that they do not overlap.

The function message-window-real-gravity should be used to compute
the proper gravity to use when displaying a message window.