Added support for pluggable completion styles.
A completion style is anything that implements
input-completion-reset and input-completion-complete.

The default completion style is now called
input-completion-style-cyclic and can be constructed
with make-input-completion-style-cyclic.

Added special variable *input-completion-style*.
Its default value is (make-input-completion-style-cyclic)
to maintain the current behaviour.

Added a new style that instead of cycling through possible completions,
completes as much as is possible unambiguously and otherwise displays
a message window with the possible completions.
This style is called input-completion-style-unambiguous.