GROUP.lisp: Miscellaneous Refactor
Extract NETWM-UPDATE-GROUP from NETWM-UPDATE-GROUPS.

ADD-GROUP: Don't include newline in the error message. Pass NAME as a
place so SBCL can provide a restart to provide a new NAME when the given
names was not valid.

GRENAME Command: Extract the main functionality of the code to %GRENAME
leaving only the precondition checks in the body of the command. The
resulting code has less nesting.