Update focus when clicking a TILE-WINDOW  behind a FLOAT-WINDOW
Previously is a floating window was 'before' a tiled window clicking
the tiled window would not focus the window. This was because the tile
window wasn't listed by TOP-WINDOWS because of an error in
WINDOW-VISIBLE-P. Fix WINDOW-VISIBLE-P so that it takes into account the
case where the tiled window is below a floating window.

Closes #548