exit keymap via specific bindings.
This adds the command call-and-exit-kmap, which allows us to run a command, and then run the exit command for the appropriate interactive keymap. It also modifies define-interactive-keymap so one can specify specific bindings that exit the keymap when run, by virtue of passing them to call-and-exit-kmap. This action is triggered by appending any true value to the end of the key binding. 
in the below example, pressing 'n' will run 'meta n', while pressing m will run 'meta x' and then exit the interactive keymap. 
(define-interactive-keymap example ()
  ((kbd "n") "meta N")
  ((kbd "m") "meta x" t))

this modification of define-interactive-keymap preserves the original usage, and doesnt break existing code using it.