Generate the heads using Xrandr instead of Xinerama.
Xinerama has several drawbacks, including that it doesn't support
dynamic reconfiguration of monitors by default (e.g. the list of
numbers is random).

Using RandR API, we notably gain the ability to get the EDID of the
monitors, which means we'll be able to store where frames were between
monitors plugs and restore them when it makes sense, for example.

Given that older machines might not support RandR (it's fairly
"young"), Xinerama support is still left in the existing
implementation.