head: use normal counters for xrandr heads.
It turns out that the EDID property returned by CLX is not unique at
all.

To fix this bug, we fallback to the previous behavior of using a
simple counter to assign the head number.

That said, because it can be useful later on, a new NAME property is
added to the head struct. _That_ property is unique, but turning it
into a number requires hashing it, which we might want to avoid as a
first step.