Make the mode-line faster and less intrusive.
Essentially, instead of redrawing the mode-line every time it is asked
to, a new thread is spawned for each mode-line, and that thread
redraws the mode-line every *mode-line-refresh-interval* (every 300ms
by default).

In practice, 300ms is not noticeable by the human eye, except for
things like the list of windows switching. For this reason, we have
another feature: whenever the mode-line is asked to be updated, we do
so, with a smaller-than-300ms throttling, defaulting to 0.04s (based
on 25 frames/sec).

Testing with this change, and the following mode-line:

```
(ql:quickload :local-time)

(defun get-date-modeline ()
  (local-time:format-timestring
   nil
   (local-time:now)
   :format '(:short-weekday #\space :short-month #\space
	     (:day 2 #\space) #\space (:hour 2) #\: (:min 2)
	     #\: (:sec 2))))

(setf *screen-mode-line-format*
      (list "^n^b %g ^n^b %W ^> "
            "  "
            "^n^b%B "
            '(:eval (get-date-modeline))))
```

I now have:

- The Slack web application no longer lags when scrolling. I believe
  it could solve the issue mentioned in #252 for a few people.
- The mode-line updates the time with the new second constantly.

I understand that this is a fairly consequential change, so I don't
expect that to be merged without discussion :-)