Fix :configuration-notify event
Before this commit stumpwm send a :configuration-notify event to the
managed xwindow that doesn't take in account the border of the
xwindow, but only the position of the parent xwindow.

This is particular important for float windows where the border is not
zero by default and there is a high top border for title.