prev next and other commands also in float-group
With this commit the commands next, prev and other are available also
in float-group.

The prev and next commands call the now generic functions
focus-prev-window and focus-next-window that specialize on the type of
the group (tile/float).

There is also a new generic function: group-windows-for-cycling
that return a list of windows in the group that can be selected. For
float-group all the windows are returned, for tile-group only the
tile-windows are returned.