new keybindings
The keybindings for "next" and "prev" command are moved
from *tile-group-root-map* to *group-root-map*. In this way they are
available also in float-group.

The key "o" now start the  command "other" in all groups. Before this
commit it started "fnext" only in tile-groups. "fnext" is still
available at TAB key in tile-groups.