Cleanup after 2cf7b54
Don't ungrab the keyboard in SHUTDOWN-INPUT-WINDOW. This used to be necessary
as SETUP-INPUT-WINDOW was responsibly for grabbing the keyboard. As of 2cf7b54
this is the responsibility of the WITH-FOCUS macro that grabs and ungrabs the
keyboard.

Remove unnecessary comments. DRAW-INPUT-BUCKET is self-explanatory and the
other comment was for a line that is no longer there.

Ref-COMMIT: 2cf7b54b92dd83b985618f141e91e2238a9599bc