Don't expose with only one application or only one frame
expose made two unconditional calls to the "only" command, which results in
spurious messages "There's only one frame." Also, it called "fselect" on only
one frame, forcing the user to select the only frame.

The new function only-one-frame-p is defined and wraps these to ensure the user
is only shown "fselect" or the error messages from "only" when it makes sense.

Fixes the bug in issue #687