replace INIT-LOAD-PATH by definition of RECURSIVELY-ADD-TO-LOAD-PATH
This new definition is also more general and can be used to add more module
directories in the user's init file. Also, this should be more save then
the original definition of INIT-LOAD-PATH.

It gets rid of the dreaded SYNC-ASDF-CENTRAL-REGISTRY that uses UNION.
This is not correct as ASDF:*CENTRAL-REGISTRY* is an ordered list and
the lisp spec says in the documentation for UNION:
  "The order of elements in the result do not have to reflect the ordering
   of list-1 or list-2 in any way."
Here is an example that shows that also SBCL's definition of UNION whould
mess with the order in SYNC-ASDF-CENTRAL-REGISTRY (as SBCL for short lists
just adds to the shorter list):
  (union (list 1 2) '(1 2 3))
  => (3 1 2)
I actually consider this behavior of SYNC-ASDF-CENTRAL-REGISTRY a bug and
this commit fixes it as the new definition uses the existing ADD-TO-LOAD-PATH
which pushes to the front of the lists.