Don't lose focus when fetching completions
There's this really annoying issue when you have a lot of completions:
while it's displaying them, the focus is lost to the underlying
window. So if you type "fi" for firefox (in `<prefix>-!`), the "i" is
sent to the underlying window.

It may be due to me using a truetype font which is slower to render,
or the fact that a single letter will yield a lot of completion
candidates.

Either way, making sure we keep the focus during the whole insertion
solves this issue.