modifies `define-frame-preference` to handle floats
This commit adds handling for a `:float` frame within
`define-frame-preference`, `remember` and `group-add-window` in order
to create a window as a floating window in a tiling group.

The keys `:match-properties-and-function` and
`:match-properties-or-function` are added in addition to *-not
variants of all properties.

A bug is fixed in `window-matches-properties-p` where type was not
properly compared.

Fixed posessive s grammar in `define-frame-preference` docstring.