Use a shell script for listing source files
Loading the ASDF Stumpwm system requires having all the Quicklisp
dependencies installed which might not be the case for a root user.
Therefore invoking a `sudo make install` is bound to fail under usual
conditions. Replacing the sources listing Lisp file with a shell script
removes the necessity to have Quicklisp dependencies installed in order
to determine Makefile dependencies.