Implemented propper cycling of windows
Implemented functions dyn-cycle-windows and dyn-rotate-windows, the
former of which wraps around the latter. dyn-rotate-windows handles
rotating the windows in the group (both stack and master) in a
clockwise or counterclockwise motion. dyn-cycle-windows handles where
focus should be when done cycling windows.

a new type is defined: :dynamic-cycle-focus. it takes one of "remain"
"master" or "follow".

dyn-cycle is rewritten to use dyn-cycle-windows.

all dynamic group commands now are only available within a dynamic
group. added keybindings for each of the dyn-cycle focus types in
the *dynamic-group-root-map*, bound to n/p N/P and M-n/p. bound o to
fnext in that map as well, to reflect tile groups.

added typechecking to swap-window-with-master.