control which window is sent to overflow group
Added variable *dynamic-group-overflow-remove-window-policy*, which
controls which window is selected to be banished from the group when
the stack frames cant be split any more. One can only select the top
or bottom of the stack as it exists BEFORE the 〈old〉 master window
is pushed onto the stack.

removed *maximum-dynamic-group-windows*, as it is irrelevant.