Changed rotation of stack, stack grows downward.
changed the semantics of rotating windows to better match the concept
of a wheel. aside from changing up/down to cl/ccl, the stack actually
gets rotated now, instead of just moving reorganizing the windows and
frames - the top or bottom window of the stack gets removed and made
into the master, and the old master gets put at the opposite end.

Changed the way frames are selected for splitting. Now the stack grows
downwards instead of outwards.