Add an idle delay to which-key-mode
This changeset makes which-key-mode more similar to emacs. Instead of popping up
immediately upon keypresses, *which-key-idle-delay* can be specified to delay
presentation of the which-key window. In addition, *which-key-secondary-idle-delay*
can be used to specify the delay for additional presentations of the which-key window
when it is already showing e.g. when navigating into another keymap.

This changeset also adds some light customization of the keybindings presentation. The
*display-bindings-formatter* parameter can be set to change presentation of
keybindings in the help window. By default, the keybindings are left unsorted to
maintain backwards compatible behavior. Two other formatting functions,
alphabetical-command-bindings-formatter and alphabetical-key-bindings-formatter
can be used to sort keybindings by command name or key sequence respectively. An
additional parameter, *which-key-bindings-formatter* can be used to adjust the
which-key keybindings format independently of the C-h format.