Allow refining input completion candidates
Completion candidates during completing-read can now be refined in a customizable
way. The *input-refine-candidates-fn* can be set to customize candidate
refinement. This parameter can be used to filter or sort candidates in novel ways.

By default, completion candidates are refined by prefix in unsorted order to maintain
backwards compatibility. Another refinement function, input-refine-regexp, is
implemented, which provides similar behavior to select-from-menu. That is, it refines
candidates by splitting the input string by space and treating each word as a regexp
to be matched.