Properly handle overflow windows for dynamic groups
Move overflow logic into its own function.

change kill-group to handle the special case of killing the overflow
group and merging its windows into a dynamic group. The previous
behaviore would move windows into the dynamic group, but if they
overflowed, they'd end up back in the overflow group. Then the group
would be killed with windows still in it. This avoids that.