Ignore XLIB:DRAWABLE-ERROR in UPDATE-CONFIGURATION.
There seems to be a problem where a window may have gone away while
updating all the windows in a group, in which case
UPDATE-CONFIGURATION fails with a synchronous error, unwinding up to
HANDLE-EVENT and aborting whatever work was being done otherwise. In
particular, this might happen during a map-request, in which case the
new window will be left incompletely set up (for example, with no
keygrabs for the TOP-MAP).

I can't say I'm sure if this is a proper way to handle the problem,
but, at least, a window that is gone shouldn't really need a
configuration-notify, so I can't see it breaking anything.