fix error where master-frame is nil if only one window is added.
Addresses an issue with dynamic groups where, when the first window is
created, if it is then destroyed, the group will try to focus the
master frame, which is nil until the group has had two windows. this
sets the master frame from the get go with the first window.