Fix moving last window of floating group to another group
When moving the window from a group to another, the window is marked as hidden, then it is acually
hidden, and group is asked to process lost focus. For float groups, the window focus lost function
picks up the next window in the group, even when it is hidden.

This patch doesn't allow a hidden window to gain focus.