No hardcoded keybindings colors and which-key format
This commit adds two new customizable variables:
- *key-seq-color*
- *which-key-format*

The former modifies the color of keybindings color as shown in the
prefix string in the which-key-mode windows.

The latter modifies how keybinds are displayed in the which-key-mode
windows.
Although *which-key-format* relies on *key-seq-color* when
initializing its value, if the latter is updated, it will be necessary
to also update the former.