Patch run-menu to only funcall functions, not remove help message
run-menu currently tries to funcall action regardless of whether its a function
or not, which conflicts with help keys returning t as the action. With this
commit we check if action is a function. additionally if we are displaying help
bindings, we dont overwrite the message window until the user presses a key, and
we allow abort keys to remove the message window instead of aborting the menu.