Additional keysym names
Add additional keysym names. This gives better readability to several
characters that can be accessed directly from the French bépo layout,
as well as the quoteleft and quoteright characters. That way, the
character itself is displayed in which-key-mode and other similar
features instead of the character’s internal name which the user
probably don’t care about that much.