Implement click support for the mode line
Add two "color" formatters, :on-click and :on-click-end, which save the bounds
of the rendered text between them to the mode line structure. These bounds are
then used by a function hung on the mode-line-click-hook which dispatches the
appropriate function with the appropriate arguments. This requires mode line
specific string renderer functions.