Add preliminary altgr support
This commit adds an altgr slot to the key structure, and defines an altgr
offset variable. The offset is used for people (such as myself) who have altgr
offsets that arent 2. A function is defined to register altgr as a modifier,
which will ensure that the function is-modifier will register altgr as a
modifier. Finally a function is defined to check if a specific keysym requires
altgr to be input.