Removed irrelevant todos, made documentation clearer
Removed a todo to change a list to an array, as the list size is small enough
that access has a negligible impact on performance.

Removed a todo to move superfluous window tracking, as it is no longer
relevant.

Changed documentation of HEAD-INFO-ALIST slot to refer to FIRST, SECOND,
etc. instead of CAR, CADR, etc.