update hooks
Introduce the function run-hook-for-minor-mode which uses the class hierarchy to
determine which hooks to run. This ensures that if a minor mode descends from
another minor mode that both of their hooks are run. They are treated similarly
to method combination in that the hooks run on enabling the minor mode are run
from least to most specific, while those run upon disabling the minor mode are
run from most to least specific.