Fix behavior of code added in #974
The initial code in pull request #974 did explicit calculations
for the modeline height, and moved the frame on y axis accordingly.
That didn't however account for people who keep their modeline in
the :BOTTOM orientation. There is also no need to manually check
and adjust for the modeline, since we can use frame-display-y and
frame-display-height instead. This is just an uneducated change,
the code should be ideally optimized and such (I don't know Lisp).