add conditions, fix nop of no-applicable-method
changed the hierarchy of conditions.

changed the no-applicable-method method for autoenable/autodisable generic
functions to signale an autoenable error, such that it wont break to debugger
but will allow handling the signal if desired.

removed the no-applicable-method for enable-when, instead having a single method
specializing on (T T) which returns nil.