Implement print-object method for swm-classes
Implement a separate way of printing swm-class instances to avoid printing the
type of the object. When minor modes are mixed in the type of the object is long
and difficult to parse, this addresses that by implementing a separate generic
function for printing swm objects and then appending the minor mode list after
them.