replace main classes properly
When replacing the root class in a dynamic mixin object, the minor modes were
not checked to see if they were compatible with the new root class. This changes
that to autodisable any minor modes whose type is not compatible with the new
root class. The filter type is not checked, as that creates a chicken/egg
situation - in order to check the filter type we need to first change the
class.