Rewrite resize-frame
Fix issues #928 and #1015.

This changes the user-facing behavior of resizing to the following:
- Find the smallest frame tree node whose split is directly below or
  to the right of the current leaf frame, if such exists. Otherwise,
  ditto for above or to the left, if such exists.
- If any such node was found, then move the split of that node
  accordingly and resize the child frame trees proportionally.

Changing the user-facing behavior is necessary to fix #1015. It will
give the user more control, but it also requires selecting the correct
frame for resizing. It also means that a resize operation can often
resize more than two frames.